+++ lib/src/templates/auth_record.dart
@@ -12,7 +12,6 @@ String authRecordClassGenerator(int lineLength) {
       ..fields.addAll([
         for (var field in authFields.where((sf) => !sf.hiddenSystem)) field.toCodeBuilder(),
       ])
-      ..methods.addAll(collectionRefMethodBuilderList.map((e) => code_builder.Method(e)))
       ..constructors.addAll([
         code_builder.Constructor((d) => d
           ..optionalParameters.addAll([