+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+* Add `fromRecordModel` factory to the generated class
+* Read `collectionId` and `collectionName` from the RecordModel
+* Set to read `json` field type as `dynamic`
+
 ## 0.0.1
 
 * Create typesafe models from `pb_schema.json`